project_a_safety_monitor/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
google-generativeai>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
